python-microservice/requirements.txt
numpy==1.26.4
opencv-python==4.8.0.76
paho-mqtt==1.6.1
pillow==10.0.1
clickhouse-driver==0.2.6
pandas==2.1.3